% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{myCatch}
\alias{myCatch}
\title{Custom Catch Function}
\usage{
myCatch(expr)
}
\arguments{
\item{expr}{Expression to evaluate.}
}
\value{
The evaluated expression.
}
\description{
Returns a data frame if the input expression causes an error, 
otherwise returns the value of the evaluated expression.
}
\keyword{internal}
