% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openPrimeRui.R
\docType{package}
\name{openPrimeRui-package}
\alias{openPrimeRui}
\alias{openPrimeRui-package}
\title{The openPrimeR Shiny Application.}
\description{
Using the openPrimeR Shiny application, you can evaluate existing
primers or design novel primers for multiplex
polymerase chain reaction that are optimized with respect
to the coverage of template sequences and the physicochemical
properties of the primers.
}
\details{
The Shiny application can be started using \code{\link{startApp}}, 
which should open the frontend in your default web browser.
}
\author{
\strong{Maintainer}: Matthias Döring \email{mdoering@mpi-inf.mpg.de}

Authors:
\itemize{
  \item Nico Pfeifer \email{pfeifer@informatik.uni-tuebingen.de}
}

}
