% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{withWarnings}
\alias{withWarnings}
\title{Error/Warning Handler.}
\usage{
withWarnings(expr)
}
\arguments{
\item{expr}{Expression to evaluate}
}
\value{
List containing values, warnings, and errors.
}
\description{
Evaluates an expression without throwing warning/errors.
}
\details{
Instead of throwing warnings/errors, the warnings/errors are
returned in a list object, alongside with the results.
}
\keyword{internal}
